feat: functional landlord/units page

+++ src/entities/building.rs
@@ -1,5 +1,6 @@
 use uuid::Uuid;
 
+#[derive(Debug, Eq, PartialEq, Hash)]
 pub struct Building {
     pub id: Uuid,
     pub landlord_id: Uuid,
@@ -11,5 +12,5 @@ pub struct BuildingSummaryRow {
     pub name: String,
     pub total_units: i64,
     pub occupied: i64,
-    pub collected: i64,
+    pub collected: i32,
 }
